Significant refactoring of websocket infrastructure
