+++ librarian_ws/responses/checksum.py
@@ -1,0 +1,7 @@
+from typing import Literal
+
+from .base import BaseResponse
+
+
+class ChecksumResponse(BaseResponse):
+    response_type: Literal["checksum"] = "checksum"
