+++ librarian_ws/responses/ping.py
@@ -1,0 +1,7 @@
+from typing import Literal
+
+from .base import BaseResponse
+
+
+class PingResponse(BaseResponse):
+    response_type: Literal["ping"] = "ping"
