+++ pyproject.toml
@@ -74,6 +74,7 @@ librarian = "hera_librarian.cli:main"
 librarian-server-rebuild-database = "librarian_server_scripts.librarian_server_rebuild_database:main"
 librarian-server-repair-database = "librarian_server_scripts.librarian_server_repair_database:main"
 librarian-change-transfer-status = "librarian_server_scripts.librarian_change_transfer_status:main"
+librarian-ws-only = "librarian_ws.serve:main"
 
 [project.urls]
 Homepage = "https://github.com/simonsobs/librarian"
